refactor: extract common files
